Minor clarity feedback
For some reason, I kept getting hung up on this one sentence.
I think it's because of its placement opposite the previous sentence
which lists two concepts (graph structure and schedules). When I first
read it, I anticipated the structure:

> One [graph structure] independently defines blah blah blah...;
> the other [schedules] yadda yadda yadda...

So when the actual meaning of the sentence unfolded differently, I had
to go back and reread it.

Happy to leave this on the cutting room floor - it's definitely a minor
nitpick and not a particularly well articulated one. Just something that
caught my eye.